requirements.txt
accelerate==0.21.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.3
catalogue==2.0.10
certifi 
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
cmake==3.29.2
colorama==0.4.6
comm==0.2.2
confection==0.1.4
ConfigArgParse==1.7
contourpy==1.2.1
cupy-cuda12x==12.1.0
cycler==0.12.1
cymem==2.0.8
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasketch==1.6.4
debugpy 
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
einops==0.6.1
einops-exts==0.0.4
en-core-web-md 
en-core-web-sm 
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.2
fastjsonschema==2.19.1
fastrlock==0.8.2
ffmpy==0.3.2
filelock==3.13.4
Flask==3.0.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.50.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.2
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
googletrans==2.4.0
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore
httplib2==0.22.0
httptools==0.6.1
httpx
huggingface-hub==0.23.0
idna==3.7
imageio==2.34.1
importlib_metadata 
importlib_resources==6.4.0
interegular==0.3.3
iopath==0.1.10
ipykernel 
ipython==8.12.3
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
joypy==0.2.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client 
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lark==1.1.9
lazy_loader==0.4
llvmlite==0.42.0
lm-format-enforcer==0.9.8
marisa-trie==1.1.0
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistralai==0.1.8
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
munkres==1.1.4
murmurhash==1.0.10
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nh3==0.2.17
ninja==1.11.1.1
nltk==3.8.1
numba==0.59.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.550.52
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf 
open3d==0.18.0
openai==1.1.1
openai-clip==1.0.1
opencv-python==4.9.0.80
orjson==3.10.1
outlines==0.0.34
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy 
peft==0.2.0
pexpect 
pickleshare 
pillow 
pipreqs==0.5.0
platformdirs==4.2.0
plotly==5.20.0
portalocker==2.8.2
preshed==3.0.9
progressbar2==4.4.2
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil 
ptyprocess 
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing 
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-utils==3.8.2
pytz 
PyYAML 
pyzmq 
ray==2.20.0
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.4.1
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.2.2
scipy==1.13.0
seaborn 
semantic-version==2.10.0
sentence-transformers==2.7.0
sentencepiece==0.1.99
sentry-sdk==2.1.1
setproctitle==1.3.3
shapely==2.0.4
shellingham==1.5.4
shortuuid==1.0.13
six 
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
statsmodels 
svgwrite==1.4.3
sympy==1.12
tenacity==8.2.3
textaugment==2.0.0
textblob==0.18.0.post0
thinc==8.2.3
threadpoolctl==3.4.0
tifffile==2024.5.3
tiktoken==0.6.0
timm==0.6.13
tinycss2==1.3.0
tokenizers==0.15.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
torchaudio
torchdata==0.7.1
torchtext
torchvision==0.16.2
tornado 
tqdm==4.66.2
traitlets==5.14.2
transformers==4.37.2
transformers-stream-generator==0.0.5
triton==2.1.0
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
unicodedata2 
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
visual-genome==1.1.1
vllm_nccl_cu12==2.18.1.0.4.0
wandb==0.17.0
wasabi==1.1.2
watchfiles==0.21.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.2
widgetsnbextension==4.0.10
xformers==0.0.23.post1
yarg==0.1.9
yarl==1.9.4
zipp 
